move `classes` from `model` to `core`
