+++ packages/core/src/classes.ts
