+++ packages/core/src/text.ts
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 
-import { Ref, Class, Doc, Obj, ArrayOf, InstanceOf, Emb, AnyLayout } from '@anticrm/model'
+import { Ref, Class, Doc, Obj, ArrayOf, InstanceOf, Emb, AnyLayout } from './classes'
 import core from '.'
 import { CreateTx, DomainIndex, Storage, Tx, TxContext } from './tx'
 import { Model } from '@anticrm/model/src/model'
