+++ packages/core/src/vdoc.ts
@@ -13,7 +13,8 @@
 // limitations under the License.
 //
 
-import { DateProperty, Doc, Emb, Ref, StringProperty, Model } from '@anticrm/model'
+import { Model } from '@anticrm/model'
+import { DateProperty, Doc, Emb, Ref, StringProperty } from './classes'
 import { CreateTx, PushTx, UpdateTx, Storage, DomainIndex, Tx, TxContext } from './tx'
 import { Space } from './space'
 import core from '.'
