+++ packages/model/src/dsl.ts
@@ -1,4 +1,3 @@
-import { RefTo, Doc, BagOf } from './classes'
 //
 // Copyright © 2020 Anticrm Platform Contributors.
 //
@@ -17,8 +16,11 @@ import { RefTo, Doc, BagOf } from './classes'
 import 'reflect-metadata'
 
 import core, {
-  Ref, Class, Obj, Mixin, ClassifierKind, Classifier, Attribute, Type, Property, mixinKey, ArrayOf, Emb, InstanceOf
+  mixinKey
 } from '.'
+import {
+  Ref, Class, Obj, Mixin, ClassifierKind, Classifier, Attribute, Type, Property, ArrayOf, Emb, InstanceOf, RefTo, Doc, BagOf
+} from '@anticrm/core'
 
 const classifierMetadataKey = Symbol("anticrm:classifier")
