+++ plugins/workbench/src/index.ts
@@ -14,12 +14,11 @@
 //
 
 import { plugin, Plugin, Service } from '@anticrm/platform'
-import { Doc, Class, Ref } from '@anticrm/model'
 
 import core from '@anticrm/platform-core'
 import ui, { AnyComponent, Asset } from '@anticrm/platform-ui'
 import { IntlString } from '@anticrm/platform-i18n'
-import { Application, VDoc } from '@anticrm/core'
+import { Application, VDoc, Doc, Class, Ref } from '@anticrm/core'
 
 export interface Perspective extends Doc {
   label: IntlString
