+++ server/server/src/__tests__/mongo_util.test.ts
@@ -16,8 +16,7 @@
 
 import { ServerSuite } from './serversuite'
 
-import { AnyLayout, BooleanProperty, StringProperty } from '@anticrm/model'
-import core from '@anticrm/core'
+import core, { AnyLayout, BooleanProperty, StringProperty } from '@anticrm/core'
 import { createSetArrayFilters } from '../mongo_utils'
 
 import { taskIds as task, createSubtask, Task, TaskComment } from '@anticrm/model/src/__tests__/test_tasks'
