+++ server/server/src/service.ts
@@ -13,13 +13,12 @@
 // limitations under the License.
 //
 
-import { AnyLayout, Class, Doc, Model, Ref } from '@anticrm/model'
 import { makeResponse, Response } from './rpc'
 import { WorkspaceProtocol } from './workspace'
 
 import { filterQuery, getUserSpaces, isAcceptable, processTx as processSpaceTx } from './spaces'
 import { Broadcaster, Client, ClientService, ClientSocket } from './server'
-import core, { Tx, SpaceUser } from '@anticrm/core'
+import core, { Tx, SpaceUser, AnyLayout, Class, Doc, Model, Ref } from '@anticrm/core'
 
 export interface ClientControl {
   ping (): Promise<void>
